+++ go/model_resp_sms_results.go
@@ -10,7 +10,7 @@ package swagger
 
 type RespSmsResults struct {
 	// Number
-	Number int32 `json:"number,omitempty"`
+	Number uint8 `json:"number,omitempty"`
 	// Phone number
 	Phone string `json:"phone,omitempty"`
 	// Some text message
